// packages/api/src/utils/supabase-types.ts
export interface definitions {
  account_keys: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    key_id: number
    name: string
    secret: string
    /**
     * Note:
     * This is a Foreign Key to `accounts.issuer`.<fk table='accounts' column='issuer'/>
     */
    issuer: string
    inserted_at: string
    updated_at: string
  }
  accounts: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    issuer: string
    sub: string
    name: string
    picture?: string
    github?: string
    email: string
    public_address: string
    inserted_at: string
    updated_at: string
  }
  contents: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    cid: string
    size?: number
    inserted_at: string
    updated_at: string
  }
  pins: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    pin_id: number
    status: 'processing' | 'queued' | 'pinning' | 'pinned' | 'failed'
    /**
     * Note:
     * This is a Foreign Key to `contents.cid`.<fk table='contents' column='cid'/>
     */
    cid: string
    service: 'PINATA' | 'IPFS_CLUSTER'
    error?: string
    inserted_at: string
    updated_at: string
  }
  uploads: {
    /**
     * Note:
     * This is a Primary Key.<pk/>
     */
    id: number
    /**
     * Note:
     * This is a Foreign Key to `accounts.issuer`.<fk table='accounts' column='issuer'/>
     */
    issuer: string
    /**
     * Note:
     * This is a Foreign Key to `account_keys.key_id`.<fk table='account_keys' column='key_id'/>
     */
    key_id?: number
    /**
     * Note:
     * This is a Foreign Key to `contents.cid`.<fk table='contents' column='cid'/>
     */
    cid?: string
    type: string
    files?: string
    inserted_at: string
    updated_at: string
  }
}